used antd pagination, added component to merge forms
